+++ subsys/bluetooth/audio/audio_internal.h
@@ -77,3 +77,5 @@ static inline const char *bt_audio_dir_str(enum bt_audio_dir dir)
 
 	return "Unknown";
 }
+
+bool bt_audio_valid_ltv(const uint8_t *data, uint8_t data_len);
