--- tests/bluetooth/audio/cap_commander/Kconfig
@@ -1,14 +1,0 @@
-# Temporary override
-#
-# Copyright (c) 2024 Nordic Semiconductor ASA
-
-# SPDX-License-Identifier: Apache-2.0
-#
-
-config BT_BAP_STREAM
-	# Override until someone fixes the UT build system
-	bool
-	default y
-
-# Include Zephyr's Kconfig.
-source "Kconfig.zephyr"
